// src/app/api/model/evento.ts
/**
 * Events API
 *
 * Contact: 97.amonsalve@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Inscripcion } from './inscripcion';


export interface Evento { 
    id?: number;
    nombre?: string | null;
    descripcion?: string | null;
    fechaHora?: string;
    ubicacion?: string | null;
    capacidadMaxima?: number;
    asistentesRegistrados?: number;
    usuarioCreadorId?: number;
    inscripciones?: Array<Inscripcion> | null;
}